Completed Android device debugging of RegisterLoginPage, DirectionsPage, and getting the SiteMapPage to load.
TODO: google maps in webview will not allow begin navigation. Maybe we need different URL parameters? Maybe we are forced to use the Maps app?

TODO: develop a way to monitor if we've reached the destination GPS coordinates.

TODO: what if we don't get a SiteFotos URL - what to display?

TODO: no more properties display; error handling with a retry action/button?

TODO: some formatting / data population in the header is missing or not quite right.